requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.2
distro==1.8.0
exceptiongroup==1.1.3
ffmpeg-python==0.2.0
future==0.18.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
numpy==1.26.2
openai==1.3.0
pydantic==2.5.1
pydantic_core==2.14.3
pydub==0.25.1
requests==2.31.0
sniffio==1.3.0
SpeechRecognition==3.10.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
whispercpp==0.0.17
